Break out the `replay-events` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`replay-events` command.
