+++ pkg/cmd/pulumi/events/replay.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package events
 
 import (
 	"encoding/json"
@@ -33,7 +33,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 )
 
-func newReplayEventsCmd() *cobra.Command {
+func NewReplayEventsCmd() *cobra.Command {
 	var preview bool
 
 	var jsonDisplay bool
